+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ jax 0.2.10 (Unreleased)
 * `GitHub commits <https://github.com/google/jax/compare/jax-v0.2.9...master>`__.
 * New features:
   * :func:`jax.scipy.stats.chi2` is now available as a distribution with logpdf and pdf methods.
+  * :func:`jax.scipy.stats.betabinom` is now available as a distribution with logpmf and pmf methods.
 
 * Bug fixes: